requirements.txt
appdirs==1.4.0
defusedxml==0.4.1
Django==1.10.5
django-allauth==0.27.0
django-braces==1.11.0
django-crispy-forms==1.6.0
django-filter==1.0.1
django-generic-positions==0.2.2
django-hvad==1.6.0
django-libs==1.67.7
django-oauth-toolkit==0.12.0
django-review==1.9.5
django-tables2==1.2.6
django-user-media==1.2.3
djangorestframework==3.5.3
easy-thumbnails==2.3
ez-setup==0.9
Markdown==2.6.7
oauthlib==2.0.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2==2.6.2
pyparsing==2.1.10
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.6.2
simplejson==3.8.2
six==1.10.0
